Move VirtualBody into submodule
